+++ PhpSerializerNET/PhpSerialization.cs
@@ -31,6 +31,9 @@ namespace PhpSerializerNET {
 		///
 		/// </returns>
 		public static object Deserialize(string input, PhpDeserializationOptions options = null) {
+			if (string.IsNullOrEmpty(input)){
+				throw new System.ArgumentException("PhpSerialization.Deserialize(): Parameter 'input' must not be null or empty.");
+			}
 			return new PhpDeserializer(input, options).Deserialize();
 		}
 
@@ -54,6 +57,9 @@ namespace PhpSerializerNET {
 			string input,
 			PhpDeserializationOptions options = null
 		) {
+			if (string.IsNullOrEmpty(input)){
+				throw new System.ArgumentException("PhpSerialization.Deserialize(): Parameter 'input' must not be null or empty.");
+			}
 			return new PhpDeserializer(input, options).Deserialize<T>();
 		}